% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{addStrandForHyperTRIBE}
\alias{addStrandForHyperTRIBE}
\title{Assign transcript strand to unstranded sites from the ADAR edit type}
\usage{
addStrandForHyperTRIBE(posGR)
}
\description{
On unstranded data a genomic A>G is an A-to-I edit on a + strand
transcript and T>C one on a - strand transcript. Sites that already carry
a strand, or have another edit type, are left unchanged.
}
\keyword{internal}
