% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotate_hits}
\alias{annotate_hits}
\title{Add gene annotations to filtered edit sites}
\usage{
annotate_hits(posGR_list_filtered, design, txi, gtfGR, ids)
}
\arguments{
\item{posGR_list_filtered}{Filtered named list of GRanges}

\item{design}{Wide-format design tibble (with control/treat cols)}

\item{txi}{tximport object}

\item{gtfGR}{Filtered GTF GRanges}

\item{ids}{Named vector gene_id -> gene_symbol}
}
\value{
Named list of annotated GRanges. Unstranded A>G sites are only
matched to + strand genes and T>C sites to - strand genes.
}
\description{
Add gene annotations to filtered edit sites
}
