% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categorise_adar}
\alias{categorise_adar}
\title{Categorise ADAR expression into low / high bins per sample}
\usage{
categorise_adar(txi, adar_tx_name = "ADARclone")
}
\arguments{
\item{txi}{tximport object}

\item{adar_tx_name}{Row name of the ADAR transgene transcript (default: "ADARclone")}
}
\value{
Named character vector ("low"/"high") with sample names
}
\description{
Categorise ADAR expression into low / high bins per sample
}
