% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{compute_feature_proportions}
\alias{compute_feature_proportions}
\title{Compute relative position of a site within its annotated feature}
\usage{
compute_feature_proportions(site, ols, feature_type, strand_val)
}
\description{
Compute relative position of a site within its annotated feature
}
\keyword{internal}
