% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_edits}
\alias{filter_edits}
\title{Filter posGR list}
\usage{
filter_edits(
  posGR_list,
  keep_edit_pairs = c("A G", "T C"),
  ep_thresh = 0.8,
  fc_thresh = 0,
  FvF = FALSE
)
}
\arguments{
\item{posGR_list}{Named list of GRanges from call_hits()}

\item{keep_edit_pairs}{Character vector of "ref targ" strings to keep (default: c("A G", "T C"))}

\item{ep_thresh}{Edit proportion threshold (default: 0.8)}

\item{fc_thresh}{Fold change threshold (default: 0)}

\item{FvF}{Logical (default: FALSE)}
}
\value{
Filtered named list of GRanges
}
\description{
Filter posGR list
}
