% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperTRIBER2-package.R
\docType{package}
\name{hyperTRIBER2-package}
\alias{hyperTRIBER2}
\alias{hyperTRIBER2-package}
\title{hyperTRIBER2: Differential RNA Editing Analysis}
\description{
Detects sites with significant differential RNA editing between conditions.
}
\author{
\strong{Maintainer}: Jackson Peter \email{jackson.peter@bio.ku.dk}

Authors:
\itemize{
  \item Jackson Peter \email{jackson.peter@bio.ku.dk}
}

}
\keyword{internal}
