% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_annotations}
\alias{load_annotations}
\title{Import and filter GTF + GFF annotations}
\usage{
load_annotations(
  gtf_file,
  gff_file,
  keep_gtf_types = c("3UTR", "5UTR", "CDS", "exon", "start_codon", "stop_codon")
)
}
\arguments{
\item{gtf_file}{Path to GTF file}

\item{gff_file}{Path to GFF3 file}

\item{keep_gtf_types}{Feature types to retain from GTF (default: c("3UTR","5UTR","CDS","exon","start_codon","stop_codon"))}
}
\value{
list(gtfGR, genes, ids)
}
\description{
Import and filter GTF + GFF annotations
}
