% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_mpileup}
\alias{load_mpileup}
\title{Load mpileup-derived text files and build matching GRanges}
\usage{
load_mpileup(mpileup_dir)
}
\arguments{
\item{mpileup_dir}{Directory containing output_*.txt files}
}
\value{
list(mpileup_df, locsGR)
}
\description{
Load mpileup-derived text files and build matching GRanges
}
