% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{lookup_reference_base}
\alias{lookup_reference_base}
\title{Look up reference bases from GRanges, handling strand complementation}
\usage{
lookup_reference_base(pos_dat, ref_gr, stranded = FALSE)
}
\description{
Look up reference bases from GRanges, handling strand complementation
}
\keyword{internal}
