% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{order_transcripts_by_expression}
\alias{order_transcripts_by_expression}
\title{Order transcripts by expression, NAs last}
\usage{
order_transcripts_by_expression(transcripts, quant)
}
\description{
Order transcripts by expression, NAs last
}
\keyword{internal}
