% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{parse_site_positions}
\alias{parse_site_positions}
\title{Parse site names into chr, pos, (strand) data frame}
\usage{
parse_site_positions(dnames, stranded = FALSE)
}
\description{
Parse site names into chr, pos, (strand) data frame
}
\keyword{internal}
