% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{resolve_gene_ambiguity}
\alias{resolve_gene_ambiguity}
\title{Resolve gene ambiguity when a site overlaps multiple genes}
\usage{
resolve_gene_ambiguity(genes, ols, pos_gr, quant, gtf_gr = ols)
}
\description{
Resolve gene ambiguity when a site overlaps multiple genes
}
\keyword{internal}
