% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_checkpoint}
\alias{save_checkpoint}
\title{Convenience wrapper: save a checkpoint .Rdat file}
\usage{
save_checkpoint(obj, res_dir, tag)
}
\arguments{
\item{obj}{Object to save}

\item{res_dir}{Directory}

\item{tag}{File name tag (e.g. "after_getHits")}
}
\description{
Convenience wrapper: save a checkpoint .Rdat file
}
