% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_hits}
\alias{call_hits}
\title{Call significant edit sites (getHits) for all experiments}
\usage{
call_hits(
  dxd_list,
  design_vectors,
  data_restricted_lists,
  locsGR,
  edits_of_interest,
  fdr = 0.1,
  symmetric = FALSE,
  ncores = 40,
  stranded = FALSE
)
}
\arguments{
\item{dxd_list}{Named list of DEXSeq results}

\item{design_vectors}{Named list of condition vectors}

\item{data_restricted_lists}{Named list from restrict_data()}

\item{locsGR}{GRanges of all mpileup positions}

\item{edits_of_interest}{Edit matrix}

\item{fdr}{FDR threshold (default: 0.1)}

\item{symmetric}{Logical (default: FALSE)}

\item{ncores}{Cores for getHits (default: 40)}

\item{stranded}{Logical (default: FALSE)}
}
\value{
Named list of GRanges with edit sites
}
\description{
Call significant edit sites (getHits) for all experiments
}
