% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{compute_site_meta}
\alias{compute_site_meta}
\title{Compute metadata for a single hit site}
\usage{
compute_site_meta(
  site_id,
  sig_res,
  data_list,
  edits_of_interest,
  design_vector,
  treat_idx,
  ctrl_idx,
  symmetric
)
}
\description{
Compute metadata for a single hit site
}
\keyword{internal}
