% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_dexseq}
\alias{run_dexseq}
\title{Run DEXSeq (make_test) for all experiments}
\usage{
run_dexseq(design_vectors, res_dir, ncores = 10)
}
\arguments{
\item{design_vectors}{Named list of named condition vectors}

\item{res_dir}{Top-level results directory}

\item{ncores}{Number of cores for DEXSeq (default: 10)}
}
\value{
Named list of DEXSeq result objects
}
\description{
Run DEXSeq (make_test) for all experiments
}
